requirements.txt
typer==0.13.0
pyfiglet==1.0.2
loguru==0.7.2
psycopg2-binary==2.9.10
faker==30.8.2
requests==2.32.3
pandas==2.2.3
httpx==0.27.2
fastapi==0.112.2
databases==0.9.0